Distributed the script into modules
